# src/Backlog.App/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Backlog",
    "description": "Agile Software",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "Quinntyne Brown",
      "email": "quinntynebrown@gmail.com"
    },
    "license": {
      "name": "Use under MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/DigitalAsset/page/{pageSize}/{index}": {
      "get": {
        "tags": [
          "DigitalAsset"
        ],
        "operationId": "GetDigitalAssetsPage",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "PageSize",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "Index",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetDigitalAssetsPageResponse"
            }
          }
        }
      }
    },
    "/api/DigitalAsset/{digitalAssetId}": {
      "get": {
        "tags": [
          "DigitalAsset"
        ],
        "operationId": "GetDigitalAssetById",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "DigitalAssetId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetDigitalAssetByIdResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "DigitalAsset"
        ],
        "operationId": "RemoveDigitalAsset",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "DigitalAssetId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/RemoveDigitalAssetResponse"
            }
          }
        }
      }
    },
    "/api/DigitalAsset/range": {
      "get": {
        "tags": [
          "DigitalAsset"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "DigitalAssetIds",
            "type": "array",
            "items": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetDigitalAssetsByIdsResponse"
            }
          }
        }
      }
    },
...
    "/api/DigitalAsset/serve/{digitalAssetId}": {
      "get": {
        "tags": [
          "DigitalAsset"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "DigitalAssetId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "format": "binary",
              "type": "string"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
...
    "/api/Profile/{profileId}": {
      "get": {
        "tags": [
          "Profile"
        ],
        "operationId": "GetProfileById",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "ProfileId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetProfileByIdResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Profile"
        ],
        "operationId": "RemoveProfile",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "ProfileId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/RemoveProfileResponse"
            }
          }
        }
      }
    },
    "/api/Profile": {
      "get": {
        "tags": [
          "Profile"
        ],
        "operationId": "GetProfiles",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetProfilesResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Profile"
        ],
        "operationId": "CreateProfile",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateProfileRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CreateProfileResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Profile"
        ],
        "operationId": "UpdateProfile",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateProfileRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UpdateProfileResponse"
            }
          }
        }
      }
    },
    "/api/Profile/page/{pageSize}/{index}": {
      "get": {
        "tags": [
          "Profile"
        ],
        "operationId": "GetProfilesPage",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "PageSize",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "Index",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetProfilesPageResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateProfileRequest": {
      "type": "object",
      "properties": {
        "profile": {
          "$ref": "#/definitions/ProfileDto"
        }
      }
    },
    "CreateProfileResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "profile": {
          "$ref": "#/definitions/ProfileDto"
        }
      }
    },
...
    "DigitalAssetDto": {
      "type": "object",
      "properties": {
        "digitalAssetId": {
          "format": "uuid",
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "bytes": {
          "format": "byte",
          "type": "string"
        },
        "contentType": {
          "type": "string"
        },
        "height": {
          "format": "float",
          "type": "number"
        },
        "width": {
          "format": "float",
          "type": "number"
        }
      }
    },
...
    "GetDigitalAssetByIdResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "digitalAsset": {
          "$ref": "#/definitions/DigitalAssetDto"
        }
      }
    },
    "GetDigitalAssetsByIdsResponse": {
      "type": "object",
      "properties": {
        "digitalAssets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DigitalAssetDto"
          }
        }
      }
    },
    "GetDigitalAssetsPageResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "length": {
          "format": "int32",
          "type": "integer"
        },
        "entities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DigitalAssetDto"
          }
        }
      }
    },
    "GetProfileByIdResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "profile": {
          "$ref": "#/definitions/ProfileDto"
        }
      }
    },
    "GetProfilesPageResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "length": {
          "format": "int32",
          "type": "integer"
        },
        "entities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProfileDto"
          }
        }
      }
    },
    "GetProfilesResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "profiles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProfileDto"
          }
        }
      }
    },
...
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    },
    "ProfileDto": {
      "type": "object",
      "properties": {
        "profileId": {
          "format": "uuid",
          "type": "string"
        },
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        }
      }
    },
...
    "RemoveDigitalAssetResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "digitalAsset": {
          "$ref": "#/definitions/DigitalAssetDto"
        }
      }
    },
    "RemoveProfileResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "profile": {
          "$ref": "#/definitions/ProfileDto"
        }
      }
    },
...
    "UpdateProfileRequest": {
      "type": "object",
      "properties": {
        "profile": {
          "$ref": "#/definitions/ProfileDto"
        }
      }
    },
    "UpdateProfileResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "profile": {
          "$ref": "#/definitions/ProfileDto"
        }
      }
    },
...
  }
}